Address PR follow-up review and CI
