+++ src/search/llm/codex.rs
@@ -167,108 +167,9 @@ fn write_file(path: &Path, content: &str) -> Result<(), String> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    #[cfg(unix)]
+    use crate::search::llm::test_support::{FakeCodex, envelope_answer_writer_script};
     use std::error::Error;
-    use std::sync::atomic::{AtomicU64, Ordering};
-
-    #[cfg(unix)]
-    static FAKE_CODEX_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-    #[cfg(unix)]
-    struct FakeCodex {
-        path: PathBuf,
-        dir: PathBuf,
-    }
-
-    #[cfg(unix)]
-    impl FakeCodex {
-        fn new(body: &str) -> Self {
-            use std::io::Write as _;
-            use std::os::unix::fs::PermissionsExt;
-
-            let dir = std::env::temp_dir().join(format!(
-                "s11-fake-codex-{}-{}",
-                std::process::id(),
-                FAKE_CODEX_COUNTER.fetch_add(1, Ordering::Relaxed)
-            ));
-            std::fs::create_dir(&dir).expect("create fake codex temp dir");
-
-            let path = dir.join("codex");
-            let mut file = std::fs::File::create(&path).expect("create fake codex script");
-            file.write_all(
-                format!(
-                    "#!/bin/sh\nif [ \"${{1:-}}\" = \"__s11_ready_probe\" ]; then exit 0; fi\nset -eu\n{}",
-                    body
-                )
-                .as_bytes(),
-            )
-            .expect("write fake codex script");
-            file.sync_all().expect("sync fake codex script");
-            drop(file);
-            let mut permissions = std::fs::metadata(&path)
-                .expect("stat fake codex script")
-                .permissions();
-            permissions.set_mode(0o700);
-            std::fs::set_permissions(&path, permissions).expect("chmod fake codex script");
-            wait_until_executable_ready(&path);
-
-            Self { path, dir }
-        }
-
-        fn path_string(&self) -> String {
-            self.path.to_string_lossy().into_owned()
-        }
-    }
-
-    #[cfg(unix)]
-    impl Drop for FakeCodex {
-        fn drop(&mut self) {
-            let _ = std::fs::remove_dir_all(&self.dir);
-        }
-    }
-
-    #[cfg(unix)]
-    fn wait_until_executable_ready(path: &Path) {
-        for _ in 0..100 {
-            match std::process::Command::new(path)
-                .arg("__s11_ready_probe")
-                .status()
-            {
-                Ok(status) if status.success() => return,
-                Ok(status) => panic!("fake codex readiness probe exited with {status}"),
-                Err(e) if e.raw_os_error() == Some(26) => std::thread::yield_now(),
-                Err(e) => panic!("fake codex readiness probe failed: {e}"),
-            }
-        }
-        panic!("fake codex executable was still busy after readiness probes");
-    }
-
-    #[cfg(unix)]
-    fn answer_writer_script(envelope: &str) -> String {
-        format!(
-            r#"answer=""
-schema=""
-while [ "$#" -gt 0 ]; do
-  case "$1" in
-    --output-schema)
-      shift
-      schema="$1"
-      ;;
-    -o)
-      shift
-      answer="$1"
-      ;;
-  esac
-  shift || true
-done
-[ -s "$schema" ]
-[ -n "$answer" ]
-cat > "$answer" <<'JSON'
-{}
-JSON
-"#,
-            envelope
-        )
-    }
 
     #[test]
     fn parses_basic_envelope() {
@@ -354,7 +255,9 @@ JSON
     #[cfg(unix)]
     #[test]
     fn invoke_codex_reads_answer_file_from_fake_cli() {
-        let fake = FakeCodex::new(&answer_writer_script(r#"{"assembly":"mov x0, x1"}"#));
+        let fake = FakeCodex::new(&envelope_answer_writer_script(
+            r#"{"assembly":"mov x0, x1"}"#,
+        ));
         let config = LlmConfig::default()
             .with_model("fake-model")
             .with_codex_bin(fake.path_string());
@@ -399,7 +302,7 @@ JSON
     #[cfg(unix)]
     #[test]
     fn invoke_codex_reports_envelope_errors() {
-        let fake = FakeCodex::new(&answer_writer_script(r#"{"assembly":"   "}"#));
+        let fake = FakeCodex::new(&envelope_answer_writer_script(r#"{"assembly":"   "}"#));
         let config = LlmConfig::default().with_codex_bin(fake.path_string());
 
         let err = invoke_codex(&config, "try a candidate", "{}").unwrap_err();
